Removes the upcase part when normalising postcode
This is covered prior, by the search form. It makes sense for that
particular behaviour to be a responsibility of the thing that performs
validation.

The trimming of whitespace remains here, this is to ensure we have
consistently formatted postcodes to add cacheability, which will be
enable soon for the production-like environments.