Add novalidate to search form.
I wanted to get client-side markup working, but doing this accessibly
was pretty out of scope at this stage. I got pretty close to it, but
it's looking like it needs a chunk more work to validate the required
checkbox group, and probably more to make it work accessibly.

So, novalidate preserves the current server-side functionality while
adding the hint to screen readers that the field is required.