Stop worrying about line/method length for spec/features/*
Solving the line/method length problem for these tests is going to be
hard and may not actually result in a better solution. For now we will
prevent these two types of 'cop' running for feature tests.

We can revisit this later. For now this means we can at least get
Rubocop into master sooner and build our new features with it's
guidance.