Fix the sharing icons in the footer
We are still using the old version of the footer. The social sharing
icons were redefined (last year) in frontend to reference white versions
instead of black ones. We need the black ones, so have added local
definitions for these for now so we can upgrade the footer in a separate
piece of effort.