Fix the bug by "Find firm directly by id"
Before the code was using searching for all firms
passing the _id as search all.

Changing to just find the specific id solves the issue.

But generates another problem. The reason that the code
does a search all is because elastic search returns a
_score attribute with the distance as the value from
where the user is located.

This commit includes the fix bug on the recently firms visited sidebar
When user search by name the app should not pass any information
of the near adviser (because the user did not provide any location).
The logic is not using the elastic feature anymore, it is relying on the
Geosort module to get the nearest adviser to show the message
on the sidebar