Fix bogus coordinates bug in profile
This commit is fixing a redundant second geocoding happening when parsing the params before sending the query to Algolia. The first one is done in the controller and is more than enough.

This bug leads to a bogus geocoding (probably in the middle of UK) when opening every firm profile, unless an actual postcode is specified.

Business requirements:

When a user searches by postcode, in the profile page you should see the closest adviser distance from the postcode.

When reaching the profile from any other type of search, you should not display any information about the closest adviser.