Change phone number displayed on firm profile page header
Currently we display the number of the first office in the list of
offices that belong to the firm. As things are indexed currently, this
is simply the most recent one that was added to the firm.

Change this so we use the firm phone number. In the index this is set to
the number of the 'main office', ie - the first office created for the
firm.

There are some potential issues here, for example the first office a
firm adds may not in fact be their main office or the primary point of
contact. Within the constraints of the current domain modelling,
however, this is the simplest change that provides the desired effect
for the majority of cases.