Update footer in line with core site
The footer on the MAS site has undergone a couple of changes
to reflect the changes in branding for the new organisation,
Money & Pensions Service. This commit changes the markup to
use the new css that upgrading yeast has pulled in and use the
new branding, specifically the new name.