Change params handling in locale switcher
Rails 5 raises an error if non-whitelisted params are used in URL
helpers. The locale switcher currently uses the entire params hash to
calculate URLs for the alternate locale.

Because this is logic in the header of the layout, it isn't feasible to
whitelist params every time these URL helpers are invoked. Instead, use
the path parameters of the request to construct a link to the
locale-switched version of the current page.

This has the side-effect of stripping all query params when switching
locale.