Fix constant resolution in LandingPage page object
This invokes FirmSection but doesn't require the file which implements
it. Tests pass locally (macos) but fail to resolve this constant on
Jenkins CI. Assuming this is due to some subtle differences between the
two platforms in how the spec support files are being required in the
spec_helper, but not entirely certain.