Update Geocoder Redis config
This library can be configured to use Redis as a cache. We have an
initializer that sets this up in production. The initializer uses the
REDISTOGO_URL environment variable, which in both our staging and
production Heroku apps is pointing to the Redis instance used by RAD
Admin. This Redis instance was replaced recently as part of the RAD
Admin Rails upgrade.

Change the env var name so that it's more obvious this relationship
exists. The Heroku app will be updated with this config prior to deploy.